Update README.md
1)Import Required Modules:

EmailMessage is imported from email. message. It is used to create an email message object.
SSL and smtplib are standard libraries for handling secure connections and sending emails, respectively.

2)Define Email Account Details:

email_sender: The email address of the sender.
email_password: The app password generated for the sender's email account. This password is used for authentication.
email_receiver: The email address of the recipient.

3)Define Email Content:

subject: The subject of the email.
body: The main content of the email.

4)Create EmailMessage Object:

An instance of EmailMessage is created and assigned to the variable em.
Email headers (From, To, Subject) and content are set for this email message.

5)Create SSL Context:

A secure SSL context is created using SSL.create_default_context().

6)Connect to SMTP Server and Send Email:

The script connects to Gmail's SMTP server (smtp.gmail.com) on port 465 using SSL.
It logs in to the sender's email account using the provided email and app password.
The sendmail method is used to send the email message to the specified recipient.

7)Print Success Message:

If the email is sent successfully, a success message is printed.